Support per provider image build settings
With mutliple cloud providers supporting different image formats and
versions nodepool needs to express the requirements providers have on
images. Add a provider.image-type and provider.image-env-vars setting to
providers that update the diskimage build settings.

image-type determines what set of image types should be emitted by
diskimage-builder when building an image.

image-env-vars are env vars set in diskimage-builders environment when
building images. These can be used to change how diskimage-builder and
its elements run as they are all shell scripts that respect env var
settings. Note that the diskimage.env-vars override any provider
image-env-vars that are set.

+++ nodepool/nodepool.py
@@ -766,14 +766,15 @@ class DiskImageBuilder(threading.Thread):
             extra_options = ('--qemu-img-options %s' %
                              image.qemu_img_options)
         img_elements = image.elements
+        img_types = image.image_types
 
         if 'fake-' in filename:
             dib_cmd = 'nodepool/tests/fake-image-create'
         else:
             dib_cmd = 'disk-image-create'
 
-        cmd = ('%s -x --no-tmpfs %s -o %s %s' %
-               (dib_cmd, extra_options, filename, img_elements))
+        cmd = ('%s -x -t %s --no-tmpfs %s -o %s %s' %
+               (dib_cmd, img_types, extra_options, filename, img_elements))
 
         log = logging.getLogger("nodepool.image.build.%s" %
                                 (image_name,))
@@ -893,10 +894,11 @@ class DiskImageUpdater(ImageUpdater):
     log = logging.getLogger("nodepool.DiskImageUpdater")
 
     def __init__(self, nodepool, provider, image, snap_image_id,
-                 filename):
+                 filename, image_type):
         super(DiskImageUpdater, self).__init__(nodepool, provider, image,
                                                snap_image_id)
         self.filename = filename
+        self.image_type = image_type
         self.image_name = image.name
 
     def updateImage(self, session):
@@ -916,10 +918,9 @@ class DiskImageUpdater(ImageUpdater):
         self.snap_image.version = timestamp
         session.commit()
 
-        # strip extension from filename
-        stripped_filename = self.filename.replace(".qcow2", "")
-        image_id = self.manager.uploadImage(image_name, stripped_filename,
-                                            'qcow2', 'bare', self.image.meta)
+        image_id = self.manager.uploadImage(image_name, self.filename,
+                                            self.image_type, 'bare',
+                                            self.image.meta)
         self.snap_image.external_id = image_id
         session.commit()
         self.log.debug("Image id: %s saving image %s" %
@@ -1232,26 +1233,6 @@ class NodePool(threading.Thread):
             g.name = g.host + '_' + str(g.port)
             newconfig.gearman_servers[g.name] = g
 
-        if 'diskimages' in config:
-            for diskimage in config['diskimages']:
-                d = DiskImage()
-                d.name = diskimage['name']
-                newconfig.diskimages[d.name] = d
-                if 'elements' in diskimage:
-                    d.elements = u' '.join(diskimage['elements'])
-                else:
-                    d.elements = ''
-                # must be a string, as it's passed as env-var to
-                # d-i-b, but might be untyped in the yaml and
-                # interpreted as a number (e.g. "21" for fedora)
-                d.release = str(diskimage.get('release', ''))
-                d.qemu_img_options = diskimage.get('qemu-img-options', '')
-                d.env_vars = diskimage.get('env-vars', {})
-                if not isinstance(d.env_vars, dict):
-                    self.log.error("%s: ignoring env-vars; "
-                                   "should be a dict" % d.name)
-                    d.env_vars = {}
-
         for provider in config['providers']:
             p = Provider()
             p.name = provider['name']
@@ -1277,6 +1258,12 @@ class NodePool(threading.Thread):
                 'template-hostname',
                 '{image.name}-{timestamp}.template.openstack.org'
             )
+            p.image_type = provider.get('image-type', 'qcow2')
+            p.image_env_vars = provider.get('image-env-vars', {})
+            if not isinstance(p.image_env_vars, dict):
+                self.log.error("%s: ignoring image-env-vars; "
+                               "should be a dict" % p.name)
+                p.image_env_vars = {}
             p.images = {}
             for image in provider['images']:
                 i = ProviderImage()
@@ -1309,6 +1296,38 @@ class NodePool(threading.Thread):
                         self.log.error("Invalid metadata for %s; ignored"
                                        % i.name)
                         i.meta = {}
+
+        if 'diskimages' in config:
+            for diskimage in config['diskimages']:
+                d = DiskImage()
+                d.name = diskimage['name']
+                newconfig.diskimages[d.name] = d
+                if 'elements' in diskimage:
+                    d.elements = u' '.join(diskimage['elements'])
+                else:
+                    d.elements = ''
+                # must be a string, as it's passed as env-var to
+                # d-i-b, but might be untyped in the yaml and
+                # interpreted as a number (e.g. "21" for fedora)
+                d.release = str(diskimage.get('release', ''))
+                d.qemu_img_options = diskimage.get('qemu-img-options', '')
+                d.env_vars = diskimage.get('env-vars', {})
+                if not isinstance(d.env_vars, dict):
+                    self.log.error("%s: ignoring env-vars; "
+                                   "should be a dict" % d.name)
+                    d.env_vars = {}
+                d.image_types = set()
+            # Do this after providers to build the image-types
+            # and qemu-img-options.
+            for provider in newconfig.providers.values():
+                for image in provider.images.values():
+                    if (image.diskimage and
+                        image.diskimage in newconfig.diskimages):
+                        diskimage = newconfig.diskimages[image.diskimage]
+                        diskimage.image_types.add(provider.image_type)
+                        env_vars = provider.image_env_vars.copy()
+                        env_vars.update(diskimage.env_vars)
+                        diskimage.env_vars = env_vars
 
         for label in config['labels']:
             l = Label()
@@ -1401,6 +1420,8 @@ class NodePool(threading.Thread):
                 new_images[k].reset != old_images[k].reset or
                 new_images[k].username != old_images[k].username or
                 new_images[k].user_home != old_images[k].user_home or
+                new_images[k].diskimage != old_images[k].diskimage or
+                new_images[k].image_type != old_images[k].image_type or
                 new_images[k].private_key != old_images[k].private_key or
                 new_images[k].meta != old_images[k].meta):
                 return False
@@ -1951,8 +1972,7 @@ class NodePool(threading.Thread):
                     self.log.debug("Queued image building task for %s" %
                                    image.name)
                     dib_image = session.createDibImage(image_name=image.name,
-                                                       filename=filename +
-                                                       ".qcow2")
+                                                       filename=filename)
 
                     # add this build to queue
                     self._image_builder_queue.put(dib_image.id)
@@ -1971,10 +1991,11 @@ class NodePool(threading.Thread):
             filename = images[0].filename
             provider_entity = self.config.providers[provider]
             provider_image = provider_entity.images[images[0].image_name]
+            image_type = provider_image.image_type
             snap_image = session.createSnapshotImage(
                 provider_name=provider, image_name=image_name)
             t = DiskImageUpdater(self, provider_entity, provider_image,
-                                 snap_image.id, filename)
+                                 snap_image.id, filename, image_type)
             t.start()
 
             # Enough time to give them different timestamps (versions)
@@ -2166,9 +2187,14 @@ class NodePool(threading.Thread):
         self.log.info("Deleted image id: %s" % snap_image.id)
 
     def deleteDibImage(self, dib_image):
+        image_config = self.config.diskimages.get(dib_image.image_name)
+        if not image_config:
+            # The config was removed deletion will have to be manual
+            return
         # Delete a dib image and it's associated file
-        if os.path.exists(dib_image.filename):
-            os.remove(dib_image.filename)
+        for image_type in image_config.image_types:
+            if os.path.exists(dib_image.filename + '.' + image_type):
+                os.remove(dib_image.filename + '.' + image_type)
 
         dib_image.state = nodedb.DELETE
         dib_image.delete()
